ModelNode into separate files
